% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{get_city_outline}
\alias{get_city_outline}
\title{Download city outline from OSM and return as sf object}
\usage{
get_city_outline(city_name)
}
\arguments{
\item{city_name}{Name of the city outline to download.}
}
\value{
\code{sf} object of downloaded data.
}
\description{
Download city outline from OSM and return as sf object
}
