% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{load_osm_data}
\alias{load_osm_data}
\title{Download OSM data, store it and and return as sf object}
\usage{
load_osm_data(bbox, crs, geom_type, key, value, clip_by_outline)
}
\arguments{
\item{bbox}{\code{vector} of corner coordinates of the area to download.}

\item{crs}{output coordinate reference system.}

\item{geom_type}{OSM geometry type.}

\item{key}{OSM key to download.}

\item{value}{OSM value to download.}

\item{clip_by_outline}{(optional) city name based on which the outline should be cut.}
}
\value{
\code{sf} object of downloaded data.
}
\description{
Wrapper function for osm_to_sf that creates a unique file name from the
passed parameters, downloads the requested data if necessary and returns the
resulting sf object.
}
\examples{
\dontrun{
# London
bb <- c (-0.90, 51.30, 0.0, 51.40)
crs <- 32630
geom_type <- "osm_points"
key <- "highway"
value <- "bus_stop"
clip_by <- "London"

bus_stops <- load_osm_data (bb, crs, geom_type, key, value, clip_by)
}
}
