% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{load_osm_value_groups}
\alias{load_osm_value_groups}
\title{Download OSM data with multiple values, combine and store them, and and
return as sf object}
\usage{
load_osm_value_groups(bbox, crs, geom_type, keys, values, group,
  clip_by_outline, timeout = 25)
}
\arguments{
\item{bbox}{\code{vector} of corner coordinates of the area to download.}

\item{crs}{output coordinate reference system.}

\item{geom_type}{OSM geometry type.}

\item{keys}{vector of OSM keys to download.}

\item{values}{vector of OSM values to download.}

\item{group}{group name for the downloaded data.}

\item{clip_by_outline}{(optional) city name based on which the outline should be cut.}

\item{timeout}{(optional) overpass timeout time in seconds.}
}
\value{
\code{sf} object of downloaded data.
}
\description{
Wrapper function for osm_to_sf that creates a unique file name from the
passed parameters, downloads the requested data if necessary and returns the
resulting sf object.
}
\examples{
\dontrun{
# London
bb <- c (-0.90, 51.30, 0.0, 51.40)
crs <- 32630
geom_type <- "osm_points"
key <- "amenity"
values <- c ("atm", "bank")
group_name <- "finance"
clip_by <- "London"

finance_london <- load_osm_value_groups (bb, crs, geom_type, key, values,
                                         group_name, clip_by)
}
}
