% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{osm_to_sf}
\alias{osm_to_sf}
\title{Download OSM data and return as sf object}
\usage{
osm_to_sf(bbox, crs, geom_type, key, value, timeout = 25)
}
\arguments{
\item{bbox}{\code{vector} of corner coordinates of the area to download.}

\item{crs}{output coordinate reference system.}

\item{geom_type}{OSM geometry type.}

\item{key}{OSM key to download.}

\item{value}{OSM value to download.}

\item{timeout}{(optional) overpass timeout time in seconds.}
}
\value{
\code{sf} object of downloaded data.
}
\description{
Download OSM data with specified parameters and return sf object.
}
\examples{
\dontrun{
# London
bb <- c (-0.90, 51.30, 0.0, 51.40)
crs <- 32630
geom_type <- "osm_points"
key <- "highway"
value <- "bus_stop"

bus_stops <- osm_to_sf (bb, crs, geom_type, key, value)
}
}
